Move builtin DataSource subtypes to module `datasources`
* Move FileDataSource/Sink and OracleDataSource/Sink to module datasources
* Move relevant tests to test_datasources
* Fix three Codacity issues
